% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DInd_fun.R
\name{DInd}
\alias{DInd}
\alias{DInd.default}
\alias{DInd.DiscreteTestResults}
\title{Discrete FWER Procedure for Independent Tests}
\usage{
DInd(test_results, ...)

\method{DInd}{default}(
  test_results,
  pCDFlist,
  alpha = 0.05,
  single_step = FALSE,
  critical_values = FALSE,
  select_threshold = 1,
  pCDFlist_indices = NULL,
  ...
)

\method{DInd}{DiscreteTestResults}(
  test_results,
  alpha = 0.05,
  single_step = FALSE,
  critical_values = FALSE,
  select_threshold = 1,
  ...
)
}
\arguments{
\item{test_results}{either a numeric vector with \eqn{p}-values or an R6 object of class \code{\link[DiscreteTests]{DiscreteTestResults}} from package \link[DiscreteTests]{DiscreteTests} for which a discrete FWER procedure is to be performed.}

\item{...}{further arguments to be passed to or from other methods. They are ignored here.}

\item{pCDFlist}{list of the supports of the CDFs of the \eqn{p}-values; each list item must be a numeric vector, which is sorted in increasing order and whose last element equals 1.}

\item{alpha}{single real number strictly between 0 and 1 indicating the target FWER level.}

\item{single_step}{single boolean specifying whether to perform a single-step (\code{TRUE}) or step-down (\code{FALSE}; the default) procedure.}

\item{critical_values}{single boolean specifying whether critical constants are to be computed.}

\item{select_threshold}{single real number strictly between 0 and 1 indicating the largest raw \eqn{p}-value to be considered, i.e. only \eqn{p}-values below this threshold are considered and the procedures are adjusted in order to take this selection effect into account; if \code{select_threshold = 1} (the default), all raw \eqn{p}-values are selected.}

\item{pCDFlist_indices}{list of numeric vectors containing the test indices that indicate to which raw \eqn{p}-value(s) each support in \code{pCDFlist} belongs; if \code{NULL} (the default) the lengths of \code{test_results} and \code{pCDFlist} \strong{must} be equal.}
}
\value{
A \code{DiscreteFWER} S3 class object whose elements are:
\item{Rejected}{rejected raw \eqn{p}-values.}
\item{Indices}{indices of rejected hypotheses.}
\item{Num_rejected}{number of rejections.}
\item{Adjusted}{adjusted \eqn{p}-values.}
\item{Critical_constants}{critical values (only exists if computations where performed with \code{critical_values = TRUE}).}
\item{Data}{list with input data.}
\item{Data$Method}{character string describing the performed algorithm, e.g. 'Discrete Bonferroni procedure'.}
\item{Data$Raw_pvalues}{observed \eqn{p}-values.}
\item{Data$pCDFlist}{list of the \eqn{p}-value supports.}
\item{Data$FWER_level}{FWER level \code{alpha}.}
\item{Data$Independence}{boolean indicating whether the \eqn{p}-values were considered as independent.}
\item{Data$Single_step}{boolean indicating whether a single-step or step-down procedure was performed.}
\item{Data$Data_name}{the respective variable names of the input data.}
\item{Select}{list with data related to \eqn{p}-value selection; only exists if \code{select_threshold < 1}.}
\item{Select$Threshold}{\eqn{p}-value selection threshold (\code{select_threshold}).}
\item{Select$Effective_Thresholds}{results of each \eqn{p}-value CDF evaluated at the selection threshold.}
\item{Select$Pvalues}{selected \eqn{p}-values that are \eqn{\leq} selection threshold.}
\item{Select$Indices}{indices of \eqn{p}-values \eqn{\leq} selection threshold.}
\item{Select$Scaled}{scaled selected \eqn{p}-values.}
\item{Select$Number}{number of selected \eqn{p}-values \eqn{\leq} selection threshold.}
}
\description{
\code{DInd()} is a wrapper function of \code{\link[=discrete_FWER]{discrete_FWER()}} for computing discrete
MTP adaptations for discrete tests. It simply passes its arguments to
\code{\link[=discrete_FWER]{discrete_FWER()}} with fixed \code{independence = TRUE}.
}
\details{
Computing critical constants (\code{critical_values = TRUE}) requires considerably
more execution time, especially if the number of unique supports is large.
We recommend that users should only have them calculated when they need them,
e.g. for illustrating the rejection area in a plot or other theoretical
reasons. Setting (\code{critical_values = FALSE}) is sufficient for obtaining
rejection decisions and adjusted \eqn{p}-values.
}
\examples{
X1 <- c(4, 2, 2, 14, 6, 9, 4, 0, 1)
X2 <- c(0, 0, 1, 3, 2, 1, 2, 2, 2)
N1 <- rep(148, 9)
N2 <- rep(132, 9)
Y1 <- N1 - X1
Y2 <- N2 - X2
df <- data.frame(X1, Y1, X2, Y2)
df

# Construction of the p-values and their supports with Fisher's exact test
library(DiscreteTests)  # for Fisher's exact test
test_results <- fisher_test_pv(df)
raw_pvalues <- test_results$get_pvalues()
pCDFlist <- test_results$get_pvalue_supports()

# d-Ind without critical values; using extracted p-values and supports
DInd_fast <- DInd(raw_pvalues, pCDFlist, single_step = TRUE)
summary(DInd_fast)

# d-Ind with critical values; using test results object
DInd_crit <- DInd(test_results, single_step = TRUE, critical_values = TRUE)
summary(DInd_crit)
 
# d-Ind (step-down) without critical values; using test results object
DInd_sd_fast <- DInd(test_results)
summary(DInd_sd_fast)

# d-Ind (step-down) with critical values; using extracted p-values and supports
DInd_sd_crit <- DInd(raw_pvalues, pCDFlist, critical_values = TRUE)
summary(DInd_sd_crit)

}
\references{
Döhler, S. (2010). Validation of credit default probabilities using
multiple-testing procedures. \emph{Journal of Risk Model Validation}, \emph{4}(4),
59-92. \doi{10.21314/JRMV.2010.062}
}
\seealso{
\code{\link[=discrete_FWER]{discrete_FWER()}}, \code{\link[=DBonf]{DBonf()}}, \code{\link[=DHolm]{DHolm()}}
}
