% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteFWER-package.R
\docType{package}
\name{DiscreteFWER-package}
\alias{DiscreteFWER}
\alias{DiscreteFWER-package}
\title{FWER-Based Multiple Testing Procedures with Adaptation for Discrete Tests}
\description{
This package implements the [d-Bonf], [sd-d-Bonf], [d-Ind], and
simulation-based procedures for discrete tests (see References).
}
\details{
Yet to come.
}
\references{
Döhler, S. (2010). Validation of credit default probabilities using
multiple-testing procedures. \emph{Journal of Risk Model Validation}, \emph{4}(4),
pp. 59-92. \doi{10.21314/JRMV.2010.062}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DISOhda/DiscreteFWER}
  \item Report bugs at \url{https://github.com/DISOhda/DiscreteFWER/issues}
}

}
\author{
\strong{Maintainer}: Florian Junge \email{diso.fbmn@h-da.de} (\href{https://orcid.org/0009-0001-6856-6938}{ORCID}) [contributor]

Authors:
\itemize{
  \item Sebastian Döhler \email{sebastian.doehler@h-da.de} (\href{https://orcid.org/0000-0002-0321-6355}{ORCID}) [contributor]
}

}
